// src/types/types.ts
export type TreeData = {
  id: string;
  treeId: string;
  areaId: string;
  treeName: string;
  achievementRate: number;
  color: string;
  branches: {
    id: string;
    name: string;
    leafs: {
      id: string;
      name: string;
      techBranch_id: string;
      techTree_id: string;
      isStatus: boolean;
    }[];
  }[];
};

//技術ツリーのみの型
export type TreeOnlyType = Pick<
  TreeInfo,
  "id" | "achievementRate" | "treeName" | "color"
>;

// 技術ブランチの型
export type TechBranch = {
  id: string;
  name: string;
  leafs: {
    id: string;
    name: string;
    techBranch_id: string;
    techTree_id: string;
    isStatus: boolean;
  }[];
};